add NamesProcessor using new base class
